+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-03T12:38:37.819Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-04T07:26:13.702Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")