+++ packages/orq-rc/RELEASES.md
@@ -10669,3 +10669,13 @@ Based on:
 - [python v4.15.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.19 - packages/orq-rc
+
+## 2026-09-04 07:29:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.20 - packages/orq-rc